requirements.txt
anthropic>=0.39.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
